requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.12.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
Django==3.2.16
django-extensions==3.2.1
djangorestframework==3.14.0
exceptiongroup==1.0.4
filelock==3.8.2
identify==2.5.10
iniconfig==1.1.1
model-bakery==1.9.0
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.0
pluggy==1.0.0
pre-commit==2.20.0
pytest==7.2.0
pytest-django==4.5.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
virtualenv==20.17.1
